% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fobi}
\alias{fobi}
\title{Food-Biomarker Ontology (.CSV) - Apr 9, 2020}
\format{
A data frame object: 1198 entities and 12 columns with different information about each entity.
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
fobi
}
\description{
This file contans all FOBI's necessary information to perform an ORA test.
}
\keyword{datasets}
