% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fobi_graph.R
\name{fobi_graph}
\alias{fobi_graph}
\title{Generate FOBI Graphs}
\usage{
fobi_graph(
  terms = NULL,
  get = NULL,
  layout = "sugiyama",
  labels = FALSE,
  labelsize = 3,
  legend = FALSE,
  legendSize = 10,
  legendPos = "bottom",
  curved = FALSE,
  pointSize = 3
)
}
\arguments{
\item{terms}{to do}

\item{get}{to do}

\item{layout}{to do}

\item{labels}{to do}

\item{labelsize}{to do}

\item{legend}{to do}

\item{legendSize}{to do}

\item{legendPos}{to do}

\item{curved}{to do}

\item{pointSize}{to do}
}
\value{
A ggraph object.
}
\description{
to do
}
\examples{

terms <- c("CHEBI:16164", "CHEBI:16243", "FOODON:00001139", "FOODON:00003274", "FOODON:00003275")

fobi_graph(terms, get = "anc")
fobi_graph(terms, get = "anc", labels = TRUE)
fobi_graph(terms, get = "anc", labels = TRUE, legend = TRUE)
fobi_graph(terms = "FOODON:00002473", labels = TRUE)

# Plot whole FOBI
fobi_graph(terms = c("FOBI:01501", "FOBI:0001"), get = "des", layout = "lgl", pointSize = 2)

}
\references{
Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.
}
\author{
Pol Castellano-Escuder
}
