% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fobi_terms}
\alias{fobi_terms}
\title{FOBI's `ontology_index` object - Feb 08, 2021}
\format{
An ontology_index object with 1201 terms.
\describe{
  \item{format-version:}{1.2}
  \item{data-version:}{fobi.owl}
  \item{ontology}{fobi}
  \item{Properties}{
  	id: character,
  	name: character,
  	parents: list,
  	children: list,
  	ancestors: list,
  	obsolete: logical}
  \item{Roots}{
  FOBI:0001 - Foods,
  FOBI:01501 - Biomarkers,
  FOBI:00422 - BiomarkerOf,
  FOBI:00423 - HasBiomarker,
  FOBI:00424 - Contains,
  FOBI:00425 - IsIngredientOf}
}
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
fobi_terms
}
\description{
FOBI terms obtained with `ontologyIndex::get_ontology()`
}
\keyword{datasets}
