% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foods}
\alias{foods}
\title{FOBI Foods Table File - Feb 02, 2021}
\format{
A tibble: 568 FOBI food entities with their associated information.
\describe{
  \item{id_code}{FOBI entity ID.}
  \item{name}{FOBI entity name.}
  \item{is_a_code}{FOBI entity direct parent class ID.}
  \item{is_a_name}{FOBI entity direct parent class name.}
  \item{id_BiomarkerOf}{FOBI entity associated food ID (in terms of biomarker).}
  \item{BiomarkerOf}{FOBI entity associated food name (in terms of biomarker).}
  \item{ChemSpider}{ChemSpider ID.}
  \item{KEGG}{KEGG ID.}
  \item{PubChemCID}{PubChemCID ID.}
  \item{InChIKey}{InChIKey ID.}
  \item{InChICode}{InChICode ID.}
  \item{FOBI}{FOBI ID.}
  \item{alias}{FOBI entity alias name.}
  \item{HMDB}{HMDB ID.}
  \item{id_Contains}{FOBI entity associated food ID (in terms of a recipe).}
  \item{Contains}{FOBI entity associated food ID (in terms of a recipe).}
}
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
foods
}
\description{
FOBI foods table obtained with `parse_fobi(terms = "FOBI:0001", get = "des")`
}
\keyword{datasets}
