% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPSrepo_foods}
\alias{GPSrepo_foods}
\title{Food-Biomarker Ontology Food Class Sets - Apr 9, 2020}
\format{
List created with `sigora::makeGPS` function.
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
GPSrepo_foods
}
\description{
This file contans FOBI's food class sets
}
\keyword{datasets}
