% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_convert.R
\name{id_convert}
\alias{id_convert}
\title{Convert Metabolite IDs Using FOBI Annotations}
\usage{
id_convert(ids, to = "FOBI", stable_version = TRUE)
}
\arguments{
\item{ids}{Character vector with metabolite IDs to convert. Input ID types can be FOBI, raw metabolite names (used in FOBI), HMDB, KEGG, PubChemCID, InChIKey, InChICode, ChemSpider or a combination of them.}

\item{to}{Target ID type. If possible, metabolites will be converted to this ID type. Options are "FOBI" (default), "metaboliteNames", "HMDB", "KEGG", "PubChemCID", "InChIKey", "InChICode" and "ChemSpider".}

\item{stable_version}{Logical. If it's set to TRUE (default), the function will use an stable version of FOBI. If not, the function will use the `fobi.obo` file from GitHub (\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}).}
}
\value{
A data frame with input IDs and converted IDs.
}
\description{
This function can convert metabolite IDs to the other ID types available in FOBI. Note that the input vector can be a combination of different IDs.
}
\author{
Pol Castellano-Escuder
}
