% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idmap}
\alias{idmap}
\title{Identifier Map File - Jan 10, 2021}
\format{
A tibble: 590 metabolites with 8 different IDs.
\describe{
  \item{metaboliteNames}{Raw metabolite names.}
  \item{FOBI}{FOBI ID.}
  \item{HMDB}{HMDB ID.}
  \item{KEGG}{KEGG ID.}
  \item{PubChemCID}{PubChemCID ID.}
  \item{InChIKey}{InChIKey ID.}
  \item{InChICode}{InChICode ID.}
  \item{ChemSpider}{ChemSpider ID.}
}
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
idmap
}
\description{
Different metabolite IDs for FOBI metabolites.
}
\keyword{datasets}
