% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_foods.R
\name{annotate_foods}
\alias{annotate_foods}
\title{Text Mining Pipeline to Annotate Free Nutritional Text into FOBI}
\usage{
annotate_foods(input, reference, similarity = 1)
}
\arguments{
\item{input}{A two column data frame. First column must contain the food ID (should be unique) and the second column must contain the food item (it can be a word or a more complex string).}

\item{reference}{Optional reference data frame with the same input structure. If this parameter is missing (default), FOBI foods will be used as a reference. To speed up the time consumption by this function a good option is to use a reference created with `fobitools::parse_fobi(terms = "FOBI:0001")`.}

\item{similarity}{Numeric between 0 (low) and 1 (high). This value indicates the semantic similarity cutoff used at the last layer of the text mining pipeline. 1 = exact match; 0 = very poor match. Values below 0.85 (aprox.) are not recomended.}
}
\value{
A data frame with the anntotated food items.
}
\description{
This function allows user to run a text mining pipeline to map nutritional free text into Food-Biomarker Ontology. This pipeline is composed of five sequential layers to map input items to FOBI food entities with the maximum accuracy as possible. See vignette for details.
}
\examples{

# Free text annotation in FOBI
free_text <- data.frame(id = 01, text = "Yesterday I ate eggs and bacon with a butter toast and black tea")
annotate_foods(free_text)

}
\author{
Pol Castellano-Escuder
}
