% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idmap}
\alias{idmap}
\title{ID Map File - Apr 9, 2020}
\format{
A data frame object: 590 metabolites with 8 possible IDs.
\describe{
  \item{metaboliteNames}{Raw metabolite names.}
  \item{FOBI}{FOBI ID.}
  \item{HMDB}{HMDB ID.}
  \item{KEGG}{KEGG ID.}
  \item{PubChemCID}{PubChemCID ID.}
  \item{InChIKey}{InChIKey ID.}
  \item{InChICode}{InChICode ID.}
  \item{ChemSpider}{ChemSpider ID.}
}
}
\source{
\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}
}
\usage{
idmap
}
\description{
This file contains all FOBI metabolites with 8 different possible IDs for each one (if possible).
}
\keyword{datasets}
