% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\title{Classical Over Representation Analysis with FOBI}
\usage{
ora(
  metabolites,
  fobi_sets = "foods",
  method = "bonferroni",
  stable_version = TRUE
)
}
\arguments{
\item{stable_version}{Logical. If it's set to TRUE (default), the function will use an stable version of FOBI. If not, the function will use the `fobi.obo` file from GitHub (\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}).}

\item{metaboliteList}{Character vector with metabolite names. Other IDs are accepted: FOBI, HMDB, KEGG, PubChemCID, InChIKey, InChICode and ChemSpider.}

\item{subOntology}{Sets desired to test for the over representation in FOBI. Options are: "foods" (default) and "chemicals".}

\item{pvalCutoff}{}

\item{adjust}{Method for multiple testing adjustment. Options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" and "none".}
}
\value{
A data frame with results.
}
\description{
This function performs a traditional Over Representation Analysis by hypergeometric test: classes are treated as sets of individual metabolites and all metabolites are treated as equally informative. `fobitools::ora` uses FOBI (Food-Biomarker Ontology) to extract the biological information and it's an adapted version of `sigora::ora` function.
}
\references{
Foroushani AB, Brinkman FS and Lynn DJ (2013). Pathway-GPS and SIGORA:identifying relevant pathways based on the over-representation of their gene-pair signatures. PeerJ, 1, pp. e229.
}
\author{
Pol Castellano-Escuder
}
