% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{ORA}
\alias{ORA}
\title{ORA}
\usage{
ORA(metabolites, method = c("bonferroni", "fdr"))
}
\arguments{
\item{metabolites}{Vector with metabolite names.}

\item{method}{Method for multiple test adjustment.}
}
\value{
A data frame with results.
}
\description{
This function performs an ORA (Over Representation Analysis) using FOBI (Food-Biomarker Ontology) given a list of metabolites. FOBIEnrich allows users to explore over represented food groups by set of metabolites.
}
\author{
Pol Castellano-Escuder
}
