% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\title{Classical Over Representation Analysis with FOBI}
\usage{
ora(
  metaboliteList,
  metaboliteUniverse,
  subOntology = "food",
  pvalCutoff = 0.01,
  adjust = "fdr"
)
}
\arguments{
\item{metaboliteList}{A vector of FOBI metabolite identifiers that define the selected list of metabolites. Other accepted identifiers are: HMDB, KEGG, PubChemCID, InChIKey, InChICode, and ChemSpider.}

\item{metaboliteUniverse}{A vector of FOBI metabolite identifiers that define the universe of possible metabolites. The identifiers should be of the same type as the `metaboliteList` identifiers.}

\item{subOntology}{A string specifying one of the two FOBI sub-ontologies: "food", or "biomarker".}

\item{pvalCutoff}{A numeric value indicating a p-value cutoff for p-values generated by the hypergeometric test.}

\item{adjust}{A string indicating a multiple testing method to use. Options are: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".}
}
\value{
A tibble with ORA results.
}
\description{
This function performs a traditional Over Representation Analysis by hypergeometric test: classes are treated as sets of individual metabolites and all metabolites are treated as equally informative. This function uses Food-Biomarker Ontology knowledge as biological information.
}
\examples{

metaboliteUniverse <- c(fobitools::idmap$FOBI[1:200], fobitools::idmap$FOBI[400:450])
metaboliteList <- c(fobitools::idmap$FOBI[1:50], fobitools::idmap$FOBI[70:80])

# Food enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               pvalCutoff = 1)

# Chemical class enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               subOntology = "biomarker", 
               pvalCutoff = 1, adjust = "fdr")

}
\references{
Foroushani AB, Brinkman FS and Lynn DJ (2013). Pathway-GPS and SIGORA:identifying relevant pathways based on the over-representation of their gene-pair signatures. PeerJ, 1, pp. e229.

Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.
}
\author{
Pol Castellano-Escuder
}
