% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fobi.R
\name{parse_fobi}
\alias{parse_fobi}
\title{Parse FOBI into Table}
\usage{
parse_fobi(terms = NULL, get = "descendants")
}
\arguments{
\item{terms}{Default is NULL. By changing this parameter for one or more entity IDs (in FOBI), this function will parse only the selected FOBI entities.}

\item{get}{Only if `terms` is not NULL. Include in the resultant table the upper (ancestors) or lower (decendants) nodes for selected entities in `terms`. Options are "descendants" (default) and "ancestors".}
}
\value{
A data frame with desired FOBI information.
}
\description{
This function parse the `fobi.obo` file from GitHub (\url{https://github.com/pcastellanoescuder/FoodBiomarkerOntology}) into readable table format.
}
\examples{

# Download and parse whole FOBI
fobi <- parse_fobi()

# Download and parse all Foods
foods <- parse_fobi(terms = "FOBI:0001")

# Download and parse all Biomarkers
biomarkers <- parse_fobi(terms = "FOBI:01501")

# Download and parse 'apple' entity and its ancestors
apple <- parse_fobi(terms = "FOODON:00002473", get = "ancestors")

}
\author{
Pol Castellano-Escuder
}
